Measurement change (#195)
* Adding option double to measurement

* nulls for status measurements

* Solving stupid datadog "no fields" approach to measurements

+++ cassandra-diagnostics-core/src/main/java/io/smartcat/cassandra/diagnostics/module/slowquery/SlowQueryModule.java
@@ -97,7 +97,7 @@ public class SlowQueryModule extends Module {
         fields.put("statement", query.statement());
 
         final Measurement measurement = Measurement
-                .create(service, query.executionTimeInMilliseconds(), query.startTimeInMilliseconds(),
+                .create(service, (double) query.executionTimeInMilliseconds(), query.startTimeInMilliseconds(),
                         TimeUnit.MILLISECONDS, tags, fields);
 
         logger.trace("Measurement transformed: {}", measurement);
